* sonarcloud fix

+++ Scripts/Items/Decorative/LowYewTable.cs
@@ -1,4 +1,3 @@
-using System;
 namespace Server.Items
 {
     public class LowYewTable : Item
@@ -28,7 +27,7 @@ namespace Server.Items
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0); // version
+            writer.Write(0); // version
         }
 
         public override void Deserialize(GenericReader reader)
